requirements.txt
streamlit
pydantic
pandas
numpy
scikit-learn
jinja2
python-docx
